Add initial implementation of Select fields using raw json and extra_params
