+++ requirements/base.txt
@@ -5,6 +5,7 @@ django-stronghold==0.3.0
 # Update djongo when release with native UUID support is released
 git+https://github.com/sshishov/djongo.git@702a96a1ce3fc128988c79a3abd744d97d561df8
 envparse==0.2.0
+jsonfield==2.0.2
 mongoengine==0.15.0
 python-logstash==0.4.6
 django-bootstrap3==9.1.0
